DataContext, seeded data, NotMapped, migarations
Added:
DataContext configuration,
Seeded data(Roles(User, Admin), Users(galkadi))

Removed:
NotMapped from User.cs and Role.cs

Error:
'Cannot create a DbSet for 'Hotel' because this type is not included in the model for the context.'

Migrations:
Migration added with (warning?):
An operation was scaffolded that may result in the loss of data. Please review the migration for accuracy.